Ngày 17/03

+++ DAY5/Toán tử/toan/toan.js
@@ -2,7 +2,7 @@ let inputWidth;
 let inputHeight;
 inputWidth = prompt("Enter the width");
 inputHeight= prompt("Enter the height");
-let Width = parseInt(inputWidth);
-let Height= parseInt(inputHeight);
+let Width = parseFloat(inputWidth);
+let Height= parseFloat(inputHeight);
 let area = Width*Height;
 document.write("The area is: "+ area)
